+++ src/settings_.rs
@@ -141,7 +141,7 @@ get_set_maker!(
 
         // mutable settings
         possible_ligands: usize {possible_ligands}, // number of possible ligands an entity can emit
-        receptors_per_entity: usize, // number of different receptor types
+        receptors_per_entity: u32, // number of different receptor types
         ligands_per_entity: u32, // number of different ligand types an entity can emit
         max_age: usize, // maximum age of an entity in simulation seconds (0 = infinite)
         max_size: f32, // maximum size of an entity
